+++ src/BindOpen.Databases.MSSqlServer/Data/Queries/Builders/DbQueryBuilder_MSSqlServer_String.cs
@@ -68,7 +68,7 @@ namespace BindOpen.Databases.Data.Queries
         /// </summary>
         /// <param name="text">The text to consider.</param>
         /// <returns>The interpreted string value.</returns>
-        public override string GetSqlText_DecodeBae64(string text)
+        public override string GetSqlText_DecodeBase64(string text)
         {
             return "decode('" + text + "', 'base64')";
         }
