+++ src/BindOpen.Databases/BindOpen.Databases.csproj
@@ -69,7 +69,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="BindOpen.Core" Version="1.1.30" />
+    <PackageReference Include="BindOpen.Core" Version="1.1.31" />
   </ItemGroup>
 
 </Project>
