+++ src/Samples.SampleB/Samples.SampleB.csproj
@@ -40,8 +40,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="BindOpen.Runtime.Logging" Version="1.1.30" />
-    <PackageReference Include="BindOpen.Runtime.NetCore" Version="1.1.30" />
+    <PackageReference Include="BindOpen.Runtime.Logging" Version="1.1.31" />
+    <PackageReference Include="BindOpen.Runtime.NetCore" Version="1.1.31" />
   </ItemGroup>
 
 </Project>
