+++ test/BindOpen.Tests.Databases/BindOpen.Tests.Databases.csproj
@@ -39,7 +39,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="BindOpen.Runtime.NetCore" Version="1.1.30" />
+    <PackageReference Include="BindOpen.Runtime.NetCore" Version="1.1.31" />
     <PackageReference Include="Bogus" Version="29.0.2" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.6.1" />
     <PackageReference Include="NUnit" Version="3.12.0" />
